deleteorder memInfo 退貨(刪除訂單)_20220617daiyun

+++ src/deleteorder.jsp
@@ -7,36 +7,43 @@
 <%
     request.setCharacterEncoding("UTF-8");
     response.setCharacterEncoding("UTF-8");
-    
-    PreparedStatement ps;
     String oid = request.getParameter("oid");
-    sql = "DELETE FROM `orders` WHERE `order_id`=?";
-    ps = con.prepareStatement(sql);
+    if(session.getAttribute("mem_account")!=null||!session.getAttribute("mem_account").equals("")){
 
-    ps.setString(1, oid);
-    int d = ps.executeUpdate();
+        PreparedStatement ps;
+        
+        sql = "DELETE FROM `orders` WHERE `order_id`=?";
+        ps = con.prepareStatement(sql);
 
-    sql = "DELETE FROM `order_details` WHERE `order_id`=?";
-    ps = con.prepareStatement(sql);
+        ps.setString(1, oid);
+        int d = ps.executeUpdate();
 
-    ps.setString(1, oid);
-    int d1 = ps.executeUpdate();
+        sql = "DELETE FROM `order_details` WHERE `order_id`=?";
+        ps = con.prepareStatement(sql);
 
-    if(d1>0&&d>0){
-        con.close();
-        %>
-        <script type="text/javascript">
-            alert("訂單刪除成功");
-            location.href="../src/memInfo.jsp";
-        </script>
-        <%
+        ps.setString(1, oid);
+        int d1 = ps.executeUpdate();
+
+        if(d1>0&&d>0){
+            con.close();
+            %>
+            <script type="text/javascript">
+                alert("訂單刪除成功");
+                location.href="../src/memInfo.jsp";
+            </script>
+            <%
+        }
+        else{
+            %>
+            <script type="text/javascript">
+                alert("訂單刪除失敗");
+                location.href="../src/memInfo.jsp";
+            </script>
+            <%
+        }
     }
     else{
-        %>
-        <script type="text/javascript">
-            alert("訂單刪除失敗");
-            location.href="../src/memInfo.jsp";
-        </script>
-        <%
+        response.sendRedirect("login.jsp");
     }
+
 %>
